Moved user code into its own class
Amended tests to work with new user class.
